Trait tor_linkspec::ChanTarget

pub trait ChanTarget: HasRelayIds + HasAddrs + HasChanMethod {
    // Provided method
    fn display_chan_target(&self) -> DisplayChanTarget<'_, Self>
       where Self: Sized { ... }
}

Information about a Tor relay used to connect to it.

Anything that implements ‘ChanTarget’ can be used as the identity of a relay for the purposes of launching a new channel.

Provided Methods

fn display_chan_target(&self) -> DisplayChanTarget<'_, Self>where Self: Sized,

Return a reference to this object suitable for formatting its ChanTarget-specific members.

The display format is not exhaustive, but tries to give enough information to identify which channel target we’re talking about.